Fix build on 32-bit targets...
This fixes commit 59a456690b3b980ef6aa9365ab78fcc47e0d42dc
which introduced a regression.

+++ src/textdialog.cpp
@@ -139,7 +139,7 @@ void TextDialog::exec() {
 				if (firstRow + rowsPerPage*2 -1 < text->size()) {
 					firstRow += rowsPerPage-1;
 				} else {
-					firstRow = max(0lu, text->size() - rowsPerPage);
+					firstRow = max(0ul, (unsigned long) (text->size() - rowsPerPage));
 				}
 				break;
 			case InputManager::SETTINGS:
